statistics: switch global stats to combined merge, thread SQLKiller
Switch the blocking and async global stats merge entrypoints from the
two-pass V1 algorithm (mergeGlobalStatsTopN -> MergePartitionHist2Global
Hist) to MergePartTopNAndHistToGlobal added in the previous code commit.

The async path now captures the caller's SQLKiller before entering
util.CallWithSCtx so KILL / connection close from the user query is
honored during the merge: CallWithSCtx hands the algorithm a pooled
stats sctx whose lifecycle is independent of the user query, so its
killer wouldn't otherwise reflect user-initiated cancellation.

Drop V1 separate-merge orchestration that the new algorithm supersedes:
- pkg/statistics/handle/globalstats/topn.go (mergeGlobalStatsTopN
  dispatcher and the parallel-by-batch path; the in-process call sites
  are gone after the switch above).
- merge_worker.go: keep StatsWrapper / NewStatsWrapper (the async path
  still ferries hist+topN through it) and drop the rest of the file.
- pkg/sessionctx/variable.SessionVars.AnalyzePartitionMergeConcurrency,
  whose only remaining reader was topn.go.
- MaxPartitionMergeBatchSize and the gpool plumbing in global_stats.go,
  unused after the switch.
- handle/util/pool.go comment update, the gpool is now described as
  serving "the global stats merge" generically rather than the deleted
  mergeGlobalStatsTopN helper.

The combined merge can end a global bucket on a value introduced by a
merge cut, which no partition bucket owns, so that bucket carries
Repeat 0. A bucket upper is a value observed in the data, so that zero
means "no point frequency recorded", not "no rows". equalRowCount would
otherwise treat it as an exact answer, so both the column and the index
path now fall through to estimateRowCountWithUniformDistribution when a
matched Repeat is zero, the same treatment
IsLastBucketEndValueUnderrepresented already applies to a matched Repeat
that looks too small.
