+++ pkg/planner/cardinality/row_count_column.go
@@ -106,8 +106,14 @@ func equalRowCountOnColumn(sctx planctx.PlanContext, c *statistics.Column, val t
 	histCnt, matched := c.Histogram.EqualRowCount(sctx, val, true)
 	// Calculate histNDV here as it's needed for both the underrepresented check and later calculations
 	histNDV := float64(c.Histogram.NDV - int64(c.TopN.Num()))
+	// A bucket's upper bound is a value observed in the data, so a zero
+	// Repeat is not a count of zero rows: it means no point frequency was
+	// recorded for it. Merged global histograms produce such buckets when
+	// an upper falls on a merge cut, and the sampled builder produces them
+	// when the estimated NDV exceeds the histogram's row count. Fall
+	// through to the uniform estimate rather than report an exact zero.
 	// also check if this last bucket end value is underrepresented
-	if matched && !IsLastBucketEndValueUnderrepresented(sctx,
+	if matched && histCnt > 0 && !IsLastBucketEndValueUnderrepresented(sctx,
 		&c.Histogram, val, histCnt, histNDV, realtimeRowCount, modifyCount) {
 		return statistics.DefaultRowEst(histCnt), nil
 	}
