+++ pkg/statistics/handle/globalstats/BUILD.bazel
@@ -5,8 +5,6 @@ go_library(
     srcs = [
         "global_stats.go",
         "global_stats_async.go",
-        "merge_worker.go",
-        "topn.go",
     ],
     importpath = "github.com/pingcap/tidb/pkg/statistics/handle/globalstats",
     visibility = ["//visibility:public"],
@@ -22,12 +20,10 @@ go_library(
         "//pkg/statistics/handle/types",
         "//pkg/statistics/handle/util",
         "//pkg/types",
-        "//pkg/util/hack",
         "//pkg/util/logutil",
         "//pkg/util/sqlkiller",
         "@com_github_pingcap_errors//:errors",
         "@com_github_pingcap_failpoint//:failpoint",
-        "@com_github_tiancaiamao_gp//:gp",
         "@org_golang_x_sync//errgroup",
         "@org_uber_go_zap//:zap",
     ],
