+++ pkg/statistics/handle/globalstats/global_stats.go
@@ -28,13 +28,7 @@ import (
 	"github.com/pingcap/tidb/pkg/statistics/handle/util"
 	"github.com/pingcap/tidb/pkg/types"
 	"github.com/pingcap/tidb/pkg/util/logutil"
-	"github.com/tiancaiamao/gp"
 	"go.uber.org/zap"
-)
-
-const (
-	// MaxPartitionMergeBatchSize indicates the max batch size for a worker to merge partition stats
-	MaxPartitionMergeBatchSize = 256
 )
 
 // statsGlobalImpl implements util.StatsGlobal
@@ -121,7 +115,7 @@ func MergePartitionStats2GlobalStats(
 		zap.Int64("tableID", globalTableInfo.ID),
 		zap.String("table", globalTableInfo.Name.L),
 	)
-	return blockingMergePartitionStats2GlobalStats(sc, statsHandle.GPool(), opts, is, globalTableInfo, isIndex, histIDs, nil, statsHandle)
+	return blockingMergePartitionStats2GlobalStats(sc, opts, is, globalTableInfo, isIndex, histIDs, nil, statsHandle)
 }
 
 // MergePartitionStats2GlobalStatsByTableID merge the partition-level stats to global-level stats based on the tableID.
@@ -167,7 +161,6 @@ func MergePartitionStats2GlobalStatsByTableID(
 // It is the old algorithm to merge partition-level stats to global-level stats. It will happen the OOM. because it will load all the partition-level stats into memory.
 func blockingMergePartitionStats2GlobalStats(
 	sc sessionctx.Context,
-	gpool *gp.Pool,
 	opts map[ast.AnalyzeOptionType]uint64,
 	is infoschema.InfoSchema,
 	globalTableInfo *model.TableInfo,
@@ -328,31 +321,27 @@ func blockingMergePartitionStats2GlobalStats(
 				return
 			}
 		}
+		allCms[i] = nil // Release for GC.
 
-		// Merge topN.
-		// Note: We need to merge TopN before merging the histogram.
-		// Because after merging TopN, some numbers will be left.
-		// These remaining topN numbers will be used as a separate bucket for later histogram merging.
-		var poppedTopN []statistics.TopNMeta
-		wrapper := NewStatsWrapper(allHg[i], allTopN[i])
-		globalStats.TopN[i], poppedTopN, allHg[i], err = mergeGlobalStatsTopN(gpool, sc, wrapper,
-			sc.GetSessionVars().StmtCtx.TimeZone(), sc.GetSessionVars().AnalyzeVersion, uint32(opts[ast.AnalyzeOptNumTopN]), isIndex)
+		// Combined TopN + histogram merge that extracts
+		// histogram upper-bound Repeat counts into the TopN counter.
+		killer := &sc.GetSessionVars().SQLKiller
+		globalStats.TopN[i], globalStats.Hg[i], err = statistics.MergePartTopNAndHistToGlobal(
+			sc.GetSessionVars().StmtCtx, killer,
+			allTopN[i], allHg[i],
+			uint32(opts[ast.AnalyzeOptNumTopN]),
+			int64(opts[ast.AnalyzeOptNumBuckets]),
+			isIndex,
+		)
+		allTopN[i] = nil // Release for GC.
+		allHg[i] = nil   // Release for GC.
 		if err != nil {
 			return
 		}
 
-		// Merge histogram.
-		globalStats.Hg[i], err = statistics.MergePartitionHist2GlobalHist(sc.GetSessionVars().StmtCtx, allHg[i], poppedTopN,
-			int64(opts[ast.AnalyzeOptNumBuckets]), isIndex, sc.GetSessionVars().AnalyzeVersion)
-		if err != nil {
-			return
-		}
-
-		// NOTICE: after merging bucket NDVs have the trend to be underestimated, so for safe we don't use them.
+		// MergePartTopNAndHistToGlobal already leaves bucket NDV = 0; here
+		// we just set the table-level NDV.
 		if globalStats.Hg[i] != nil {
-			for j := range globalStats.Hg[i].Buckets {
-				globalStats.Hg[i].Buckets[j].NDV = 0
-			}
 			globalStats.Hg[i].NDV = globalStatsNDV
 		}
 	}
