+++ pkg/statistics/handle/globalstats/global_stats_async.go
@@ -33,6 +33,7 @@ import (
 	statstypes "github.com/pingcap/tidb/pkg/statistics/handle/types"
 	"github.com/pingcap/tidb/pkg/statistics/handle/util"
 	"github.com/pingcap/tidb/pkg/types"
+	"github.com/pingcap/tidb/pkg/util/sqlkiller"
 	"go.uber.org/zap"
 	"golang.org/x/sync/errgroup"
 )
@@ -45,6 +46,20 @@ type mergeItem[T any] struct {
 type skipItem struct {
 	histID      int64
 	partitionID int64
+}
+
+// StatsWrapper wrapper stats
+type StatsWrapper struct {
+	AllHg   []*statistics.Histogram
+	AllTopN []*statistics.TopN
+}
+
+// NewStatsWrapper returns wrapper
+func NewStatsWrapper(hg []*statistics.Histogram, topN []*statistics.TopN) *StatsWrapper {
+	return &StatsWrapper{
+		AllHg:   hg,
+		AllTopN: topN,
+	}
 }
 
 // toSQLIndex is used to convert bool to int64.
@@ -261,7 +276,7 @@ func (a *AsyncMergePartitionStats2GlobalStats) ioWorker(sctx sessionctx.Context,
 	return nil
 }
 
-func (a *AsyncMergePartitionStats2GlobalStats) cpuWorker(stmtCtx *stmtctx.StatementContext, sctx sessionctx.Context, opts map[ast.AnalyzeOptionType]uint64, isIndex bool, tz *time.Location, analyzeVersion int) (err error) {
+func (a *AsyncMergePartitionStats2GlobalStats) cpuWorker(stmtCtx *stmtctx.StatementContext, killer *sqlkiller.SQLKiller, opts map[ast.AnalyzeOptionType]uint64, isIndex bool) (err error) {
 	defer func() {
 		if r := recover(); r != nil {
 			statslogutil.StatsLogger().Warn("cpuWorker panic", zap.Stack("stack"), zap.Any("error", r))
@@ -292,7 +307,7 @@ func (a *AsyncMergePartitionStats2GlobalStats) cpuWorker(stmtCtx *stmtctx.Statem
 			panic("test for PanicSameTime")
 		}
 	})
-	err = a.dealHistogramAndTopN(stmtCtx, sctx, opts, isIndex, tz, analyzeVersion)
+	err = a.dealHistogramAndTopN(stmtCtx, killer, opts, isIndex)
 	if err != nil {
 		statslogutil.StatsLogger().Warn("dealHistogramAndTopN failed", zap.Error(err))
 		return err
@@ -312,9 +327,12 @@ func (a *AsyncMergePartitionStats2GlobalStats) MergePartitionStats2GlobalStats(
 	isIndex bool,
 ) error {
 	a.skipMissingPartitionStats = sctx.GetSessionVars().SkipMissingPartitionStats
-	tz := sctx.GetSessionVars().StmtCtx.TimeZone()
-	analyzeVersion := sctx.GetSessionVars().AnalyzeVersion
 	stmtCtx := sctx.GetSessionVars().StmtCtx
+	// Capture the caller's SQLKiller here. Inside CallWithSCtx the
+	// passed-in sctx is a pooled stats session whose lifecycle is
+	// independent of the user's query, so its killer wouldn't reflect
+	// user-initiated KILL / connection close.
+	killer := &sctx.GetSessionVars().SQLKiller
 	return util.CallWithSCtx(a.statsHandle.SPool(),
 		func(sctx sessionctx.Context) error {
 			err := a.prepare(sctx, isIndex)
@@ -328,7 +346,7 @@ func (a *AsyncMergePartitionStats2GlobalStats) MergePartitionStats2GlobalStats(
 				return a.ioWorker(sctx, isIndex)
 			})
 			mergeWg.Go(func() error {
-				return a.cpuWorker(stmtCtx, sctx, opts, isIndex, tz, analyzeVersion)
+				return a.cpuWorker(stmtCtx, killer, opts, isIndex)
 			})
 			err = metawg.Wait()
 			if err != nil {
@@ -490,7 +508,7 @@ func (a *AsyncMergePartitionStats2GlobalStats) dealCMSketch() error {
 	}
 }
 
-func (a *AsyncMergePartitionStats2GlobalStats) dealHistogramAndTopN(stmtCtx *stmtctx.StatementContext, sctx sessionctx.Context, opts map[ast.AnalyzeOptionType]uint64, isIndex bool, tz *time.Location, analyzeVersion int) (err error) {
+func (a *AsyncMergePartitionStats2GlobalStats) dealHistogramAndTopN(stmtCtx *stmtctx.StatementContext, killer *sqlkiller.SQLKiller, opts map[ast.AnalyzeOptionType]uint64, isIndex bool) (err error) {
 	failpoint.Inject("dealHistogramAndTopNErr", func(val failpoint.Value) {
 		if val, _ := val.(bool); val {
 			failpoint.Return(errors.New("dealHistogramAndTopNErr returned error"))
@@ -509,28 +527,23 @@ func (a *AsyncMergePartitionStats2GlobalStats) dealHistogramAndTopN(stmtCtx *stm
 				return nil
 			}
 			var err error
-			var poppedTopN []statistics.TopNMeta
-			var allhg []*statistics.Histogram
+			// Combined TopN + histogram merge.
 			wrapper := item.item
-			a.globalStats.TopN[item.idx], poppedTopN, allhg, err = mergeGlobalStatsTopN(a.statsHandle.GPool(), sctx, wrapper,
-				tz, analyzeVersion, uint32(opts[ast.AnalyzeOptNumTopN]), isIndex)
-			if err != nil {
-				return err
-			}
-
-			// Merge histogram.
 			globalHg := &(a.globalStats.Hg[item.idx])
-			*globalHg, err = statistics.MergePartitionHist2GlobalHist(stmtCtx, allhg, poppedTopN,
-				int64(opts[ast.AnalyzeOptNumBuckets]), isIndex, analyzeVersion)
+			a.globalStats.TopN[item.idx], *globalHg, err = statistics.MergePartTopNAndHistToGlobal(
+				stmtCtx, killer,
+				wrapper.AllTopN, wrapper.AllHg,
+				uint32(opts[ast.AnalyzeOptNumTopN]),
+				int64(opts[ast.AnalyzeOptNumBuckets]),
+				isIndex,
+			)
 			if err != nil {
 				return err
 			}
 
-			// NOTICE: after merging bucket NDVs have the trend to be underestimated, so for safe we don't use them.
+			// MergePartTopNAndHistToGlobal already leaves bucket NDV = 0; here
+			// we just set the table-level NDV.
 			if *globalHg != nil {
-				for j := range (*globalHg).Buckets {
-					(*globalHg).Buckets[j].NDV = 0
-				}
 				(*globalHg).NDV = a.globalStatsNDV[item.idx]
 			}
 		case <-a.ioWorkerExitWhenErrChan:
